Все-таки нужно два разных радиотапа

+++ libwfb/include/private/radiotap.h
@@ -20,21 +20,18 @@
 
 struct ieee80211_radiotap_iterator {
 	struct ieee80211_radiotap_header *rtheader;
-	int max_length;
+	size_t max_length;
 	int this_arg_index;
 	uint8_t *this_arg;
 
-	int arg_index;
+	size_t arg_index;
 	uint8_t *arg;
 	__le32 *next_bitmap;
 	uint32_t bitmap_shifter;
 };
 
-extern int ieee80211_radiotap_iterator_init(
-   struct ieee80211_radiotap_iterator *iterator,
-   struct ieee80211_radiotap_header *radiotap_header,
-   int max_length);
+extern int ieee80211_radiotap_iterator_init(struct ieee80211_radiotap_iterator *iterator,
+					    struct ieee80211_radiotap_header *radiotap_header,
+					    int max_length);
 
-extern int ieee80211_radiotap_iterator_next(
-   struct ieee80211_radiotap_iterator *iterator);
-
+extern int ieee80211_radiotap_iterator_next(struct ieee80211_radiotap_iterator *iterator);
